# server/docs/swagger_output_dev_new.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/v1/student/event/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "eventName": {
                  "example": "any"
                },
                "eventDescription": {
                  "example": "any"
                },
                "eventDate": {
                  "example": "any"
                },
                "eventTime": {
                  "example": "any"
                },
                "eventVenue": {
                  "example": "any"
                },
                "eventOrganiser": {
                  "example": "any"
                },
                "participantCount": {
                  "example": "any"
                },
                "hostedBy": {
                  "example": "any"
                },
                "eventPoster": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 4
                    },
                    "eventName": {
                      "type": "string",
                      "example": "WebVerse3.0"
                    },
                    "eventDescription": {
                      "type": "string",
                      "example": "fiuhgudasisnlgfdg"
                    },
                    "eventPoster": {
                      "type": "string",
                      "example": "url://123/img.jpg"
                    },
                    "eventDate": {
                      "type": "string",
                      "example": "10/10/2023"
                    },
                    "eventTime": {
                      "type": "string",
                      "example": "10:00"
                    },
                    "eventVenue": {
                      "type": "string",
                      "example": "MG Audi"
                    },
                    "eventOrganiser": {
                      "type": "string",
                      "example": "Android Club"
                    },
                    "participantCount": {
                      "type": "number",
                      "example": 12
                    },
                    "hostedBy": {
                      "type": "string",
                      "example": "Android Club"
                    },
                    "isApproved": {
                      "type": "boolean",
                      "example": false
                    },
                    "isRejected": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 4
                      },
                      "eventName": {
                        "type": "string",
                        "example": "WebVerse3.0"
                      },
                      "eventDescription": {
                        "type": "string",
                        "example": "fiuhgudasisnlgfdg"
                      },
                      "eventPoster": {
                        "type": "string",
                        "example": "url://123/img.jpg"
                      },
                      "eventDate": {
                        "type": "string",
                        "example": "10/10/2023"
                      },
                      "eventTime": {
                        "type": "string",
                        "example": "10:00"
                      },
                      "eventVenue": {
                        "type": "string",
                        "example": "MG Audi"
                      },
                      "eventOrganiser": {
                        "type": "string",
                        "example": "Android Club"
                      },
                      "participantCount": {
                        "type": "number",
                        "example": 12
                      },
                      "hostedBy": {
                        "type": "string",
                        "example": "Android Club"
                      },
                      "isApproved": {
                        "type": "boolean",
                        "example": true
                      },
                      "isRejected": {
                        "type": "boolean",
                        "example": false
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/v1/warden/student/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "John Dow"
                    },
                    "regNo": {
                      "type": "string",
                      "example": "21BCE1000"
                    },
                    "block": {
                      "type": "string",
                      "example": "A"
                    },
                    "roomNo": {
                      "type": "string",
                      "example": "A-102"
                    },
                    "messType": {
                      "type": "string",
                      "example": "Special"
                    },
                    "toBeChangedTo": {
                      "type": "string",
                      "example": ""
                    },
                    "isForChange": {
                      "type": "boolean",
                      "example": false
                    },
                    "roomId": {
                      "type": "number",
                      "example": 3
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/warden/student/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 2
                    },
                    "name": {
                      "type": "string",
                      "example": "John Wick"
                    },
                    "regNo": {
                      "type": "string",
                      "example": "21BPS1388"
                    },
                    "block": {
                      "type": "string",
                      "example": "A"
                    },
                    "roomNo": {
                      "type": "string",
                      "example": ""
                    },
                    "messType": {
                      "type": "string",
                      "example": "Veg"
                    },
                    "toBeChangedTo": {
                      "type": "string",
                      "example": ""
                    },
                    "isForChange": {
                      "type": "boolean",
                      "example": false
                    },
                    "roomId": {}
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}